initializeAnswer()、updateAnswerCount()を共通化
