+++ app/javascript/new-answer.js
@@ -2,7 +2,8 @@ import CSRF from 'csrf'
 import TextareaInitializer from 'textarea-initializer'
 import MarkdownInitializer from 'markdown-initializer'
 import { toast } from './vanillaToast.js'
-import { initializeAnswer, updateAnswerCount } from './answer.js'
+import updateAnswerCount from './updateAnswerCount.js'
+import initializeAnswer from './initializeAnswer.js'
 import { initializeReaction } from './reaction.js'
 import store from './check-store.js'
